Javadoc changes
