+++ src/main/java/frc/robot/filters/Filter.java
@@ -1,5 +1,10 @@
 package frc.robot.filters;
 
 public interface Filter {
+    /**
+     * Calculate the next output from the input.
+     * @param input The input variable
+     * @return The output of the filter
+     */
     public double calculate(double input);
 }
